add autpi sim APN and removed link to comming soon
I've added a note to the begginers guide hte note that the APN string should be nxt20.net if you have recvied a device with a preinstalled SIM card. I have also removed the create_can_signal_loggers from the sidebars as there's no information in there, and just says "comming soon" and has done so for a while.